+++ packages/hardhat-graph-protocol/CHANGELOG.md
@@ -1,5 +1,12 @@
 # hardhat-graph-protocol
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @graphprotocol/toolshed@0.5.0
+
 ## 0.1.15
 
 ### Patch Changes